# photo-prestige/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Photo Prestige API Gateway",
    "version": "1.2.0",
    "description": "Complete gateway documentation for all available API endpoints.\\n\\nRole model: participant, target-owner.\\n\\nEach operation includes role access in the description and x-allowed-roles."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local gateway"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "TargetCreate": {
        "type": "object",
        "required": [
          "title",
          "imageUrl",
          "lat",
          "lng",
          "deadlineAt"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri"
          },
          "locationDescription": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          },
          "radiusMeters": {
            "type": "number",
            "example": 250
          },
          "deadlineAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DeadlinePatch": {
        "type": "object",
        "required": [
          "deadlineAt"
        ],
        "properties": {
          "deadlineAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/targets": {
      "get": {
        "tags": [
          "Targets"
        ],
        "summary": "List targets",
        "description": "Allowed roles: public.",
        "x-allowed-roles": [
          "public"
        ],
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ownerId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "lat",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radiusMeters",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radiusKm",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Targets list"
          },
          "400": {
            "description": "Invalid query"
          }
        }
      },
      "post": {
        "tags": [
          "Targets"
        ],
        "summary": "Create target",
        "description": "Allowed roles: target-owner.",
        "x-allowed-roles": [
          "target-owner"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Target created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/targets/{targetId}": {
      "get": {
        "tags": [
          "Targets"
        ],
        "summary": "Get target by id",
        "description": "Allowed roles: public.",
        "x-allowed-roles": [
          "public"
        ],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Target details"
          },
          "404": {
            "description": "Target not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Targets"
        ],
        "summary": "Delete target",
        "description": "Allowed roles: target-owner (must own target).",
        "x-allowed-roles": [
          "target-owner"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Target deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Target not found"
          }
        }
      }
    },
    "/api/v1/targets/{targetId}/deadline": {
      "patch": {
        "tags": [
          "Targets"
        ],
        "summary": "Update target deadline",
        "description": "Allowed roles: target-owner (must own target).",
        "x-allowed-roles": [
          "target-owner"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeadlinePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deadline updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Target not found"
          }
        }
      }
    },
...
  }
}